src/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "email-validator>=2.2.0",
    "fastapi>=0.115.6",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.5",
    "redis>=5.2.1",
    "uvicorn>=0.34.0",
]
